finished break up of record exercise view
